currently in-progress code. kills savefiles. squads follow
Currently working code. unforunately i can only save/load the game once, maybe twice. eventually i get  crashes when trying to load a save with the droids in it. I haven't tried the case where all droids are dead though. The debugging text has been removed, it is playable, just dont expect the save to work afterwards. 

On a player-by-player basis, a squad and soldier table with their name as the key is created, and maintained. The squad will now follow the player around. this feature works fully, currently.  Next step is to make stationary squads by creating a new group and copying all the current squad info over and clearing the current tables, and later should have a way to  re-absorb that stationary  squad.  I'm thinking the use of planner selection boxes or something.