some more cleanup, mostly around how trimSquad and checkMembers (now renamed to validateSquadIntegrity) work.
the basic idea is to try to do as few checks as possible, in the proper locations.

also, though, the updated functions either return the squad itself, once
trimmed or validated, or nil, to indicate failure.
This is a move toward a slightly more "pure functional" style codebase,
which should go a very long way towards keeping the code maintainable.

lastly, removed the squad.members.size attribute in favor of a new
squad.numMembers attribute that gets updated inside trimSquad.