added simple assembler-centricity option.
Basically assembler-centricity just means that instead of looking for the
enemy closest to the squad's current position, it first looks for the closest assembler,
and then looks for the enemy closest to that assembler.

This enables players to use their assembler placement as a sort of
"kill radius designation". It encourages players to actively "plan"
what areas they want to be kept clear.

Currently it is enabled by default, but this probably shouldn't be the case?

This is something I would be interested in adding to the options in the DroidSettingsModule,
but it looks like we'd need to create a graphic with the text for it to make it work.

----

additionally, threw in some changes that will help retreating squads respond better
to changes in where assemblers are placed.

----

Lastly, in the name of player control, I have added a check that will
cause squads to request new orders if the player manually adds a unit to
the squad. All squads will always eventually request new orders, but this
gives the player at least some minor amount of control.
An example of where this might be useful is if the player has recently
placed an assembler that would be a closer retreat option for a retreating squad,
or is using assembler centricity and wants the squad to attack a different location.
In prior versions of Robot Army, the player would simply have to wait for the
active command to finish. Now, manually placing even a single soldier next to the squad
will cause the squad to re-evaluate its mission within the next tick.